Updated SyncThing package but then realized it didn't work on one of my 32-bit tablets... Also added additional statement regarding upgrading between major versions.

+++ syncthing/tools/chocolateyInstall.ps1
@@ -8,40 +8,46 @@ if ($psISE) {
     Import-Module -name "$env:ChocolateyInstall\chocolateyinstall\helpers\chocolateyInstaller.psm1"
 }
 
-try {
-    if (-not (Test-Path "$appDir\syncthing.exe")) {
-        if (-not (Test-Path $downloadPath)) {
-            New-Item -Type Directory -Path $downloadPath | Out-Null
-        }
-
-        Get-ChocolateyWebFile $packageName "$downloadPath\$packageName.zip" $url $url64
-        Get-ChocolateyUnzip "$downloadPath\$packageName.zip" "$downloadPath\"
-
-        if (Test-Path $appDir) {
-            Write-Output "Removing previous version of package..."
-            Remove-Item -Path $appDir -Recurse -Force
-        }
-
-        New-Item -Type Directory -Path $appDir | Out-Null
-
-        Push-Location "$downloadPath\syncthing-windows-amd64-v*"
-
-        Copy-Item -Path "*" -Destination "$appDir\" -Recurse -Container
-
-        Pop-Location
-
-        $location = "HKCU:\Software\Microsoft\Windows\CurrentVersion\Run"
-
-        $key = Get-Item $location
-        if ($key.GetValue($packageName, $null) -ne $null) {
-            Remove-ItemProperty -Path $location -Name $packageName
-        }
-
-        New-ItemProperty -Path $location -Name $packageName -Value "$appdir\syncthing.exe -no-console -no-browser"
+if (-not (Test-Path "$appDir\syncthing.exe")) {
+    if (-not (Test-Path $downloadPath)) {
+        New-Item -Type Directory -Path $downloadPath | Out-Null
     }
 
-    Write-ChocolateySuccess $packageName
-} catch {
-    Write-ChocolateyFailure $packageName $($_.Exception.Message)
-    throw
+    Get-ChocolateyWebFile $packageName "$downloadPath\$packageName.zip" $url $url64
+    Get-ChocolateyUnzip "$downloadPath\$packageName.zip" "$downloadPath\"
+
+    if (Test-Path $appDir) {
+        Write-Output "Removing previous version of package..."
+        Remove-Item -Path $appDir -Recurse -Force
+    }
+
+    New-Item -Type Directory -Path $appDir | Out-Null
+
+    if (Test-Path "$downloadPath\syncthing-windows-amd64-v*") {
+        Push-Location "$downloadPath\syncthing-windows-amd64-v*"
+    }
+
+    if (Test-Path "$downloadPath\syncthing-windows-386-v*") {
+        Push-Location "$downloadPath\syncthing-windows-386-v*"
+    }
+
+    if (Test-Path ".\syncthing.exe" ) {
+        Copy-Item -Path "*" -Destination "$appDir\" -Recurse -Container
+    } else {
+        throw "Syncthing executable not found! Did the download fail?"
+    }
+
+    Pop-Location
+
+    $location = "HKCU:\Software\Microsoft\Windows\CurrentVersion\Run"
+
+    $key = Get-Item $location
+    if ($key.GetValue($packageName, $null) -ne $null) {
+        Remove-ItemProperty -Path $location -Name $packageName
+    }
+
+    New-ItemProperty -Path $location -Name $packageName -Value "$appdir\syncthing.exe -no-console -no-browser"
+} else {
+    Write-Warning "Syncthing is already installed and should self-update."
+    Write-Output "If you are trying to update between major versions, uninstall previous version first..."
 }
